separating models for the two databases
